+++ lib/movesdk.dart
@@ -3,12 +3,15 @@ import 'package:movesdk/io/dolphin/move/move_auth.dart';
 import 'package:movesdk/io/dolphin/move/move_auth_error.dart';
 import 'package:movesdk/io/dolphin/move/move_auth_state.dart';
 import 'package:movesdk/io/dolphin/move/move_detection_service.dart';
+import 'package:movesdk/io/dolphin/move/move_device.dart';
 import 'package:movesdk/io/dolphin/move/move_geocode_result.dart';
+import 'package:movesdk/io/dolphin/move/move_scan_result.dart';
 import 'package:movesdk/io/dolphin/move/move_service_warning.dart';
 import 'package:movesdk/io/dolphin/move/move_shutdown_result.dart';
 import 'package:movesdk/io/dolphin/move/move_state.dart';
 import 'package:movesdk/io/dolphin/move/move_trip_state.dart';
 
+import 'io/dolphin/move/move_options.dart';
 import 'movesdk_platform_interface.dart';
 
 /// A list of services passed in `setup(auth, config)` or `updateConfig(config)`.
@@ -42,8 +45,8 @@ class MoveSdk {
   /// [moveAuth] contains authentication data and tokens prepared by the app backend..
   /// [moveConfig] indicates the configuration of the services which will be running.
   /// Services in [moveConfig] must be enabled in the MOVE dashboard.
-  Future<void> setup(MoveAuth auth, MoveConfig moveConfig) {
-    return MovesdkPlatform.instance.setup(auth, moveConfig);
+  Future<void> setup(MoveAuth auth, MoveConfig moveConfig, {MoveOptions? options}) {
+    return MovesdkPlatform.instance.setup(auth, moveConfig, options);
   }
 
   /// Get a unique Device Identifier to distinguish the device.
@@ -276,5 +279,38 @@ class MoveSdk {
   /// or permission errors.
   Stream<List<MoveServiceError>> setServiceErrorListener() async* {
     yield* MovesdkPlatform.instance.setServiceErrorListener();
+  }
+
+  /// Starts scanning for devices that can be registered with the sdk
+  /// Scan can be filtered with [filter], default includes only paired devices.
+  /// For scanning beaons [uuid] and [manufacturerId] must be specified.
+  /// Will stop when stream is closed.
+  Stream<List<MoveDevice>> startScanningDevices(
+      {List<MoveDeviceFilter> filter = const [MoveDeviceFilter.paired],
+      String? uuid,
+      int? manufacturerId}) async* {
+    yield* MovesdkPlatform.instance
+        .startScanningDevices(filter: filter, uuid: uuid, manufacturerId: manufacturerId);
+  }
+
+  /// Get a list of devices registered with the sdk to be scanned for during trip.
+  Future<List<MoveDevice>> getRegisteredDevices() {
+    return MovesdkPlatform.instance.getRegisteredDevices();
+  }
+
+  /// Register devices with the sdk to be scanned for during trip.
+  /// All will be unregistered on shutdown.
+  Future<void> registerDevices(List<MoveDevice> devices) {
+    return MovesdkPlatform.instance.registerDevices(devices);
+  }
+
+  /// Unregister devices with the sdk to be scanned for during trip
+  Future<void> unregisterDevices(List<MoveDevice> devices) {
+    return MovesdkPlatform.instance.unregisterDevices(devices);
+  }
+
+  /// Device listener fired on device scans during trips.
+  Stream<List<MoveScanResult>> setDeviceDiscoveryListener() async* {
+    yield* MovesdkPlatform.instance.setDeviceDiscoveryListener();
   }
 }